+++ src/expression/Expression.java
@@ -1,5 +1,5 @@
 package expression;
 
 public interface Expression {
-	public int calculate(String[] args, int index) throws Exception;
+	public double calculate(String[] args, int index) throws Exception;
 }
